+++ src/actions/index.js
@@ -1,4 +1,5 @@
 export * from "./address";
+export * from "./analytics";
 export * from "./app";
 export * from "./authentication";
 export * from "./city";